Basic functionality and logging

+++ copychanged/Program.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
+using System.Text;
 using System.Threading;
 
 namespace copychanged
@@ -10,6 +11,8 @@ namespace copychanged
     class FileToCopy {
         public string from;
         public string to;
+        public UInt64 size;
+        public UInt64 sizeTarget;
         public bool existsAndDifferent;
     }
 
@@ -23,6 +26,8 @@ namespace copychanged
             Console.WriteLine("Optional param meaning:");
             Console.WriteLine("-n,--donew        Copy files that didn't exist yet.");
             Console.WriteLine("-c,--dochanged    Copy files that are different (overwrite!).");
+            Console.WriteLine("-l,--log          Log what was done into _copychanged_copy.log and _copychanged_fix.log.");
+            Console.WriteLine("-v,--verbose      Logging will include list of files that need to be copied (not shown by default to avoid spam). By default, _copychanged_copy.log only contains any errors or retries during copying and the count of files.");
             Console.WriteLine("By default you are only told the amount of needed copy operations, no copying happens.");
             Console.WriteLine("Copy operations are ALWAYS verified and will be repeated if a difference is found.");
         }
@@ -37,6 +42,8 @@ namespace copychanged
 
             bool doNew = false;
             bool doChanged = false;
+            bool doLog = false;
+            bool verboseLog = false;
 
             string folder1 = null;
             string folder2 = null;
@@ -52,6 +59,16 @@ namespace copychanged
                 {
                     Console.WriteLine("--dochanged/-c option detected. Will overwrite files that changed.");
                     doChanged = true;
+                } 
+                else if (arg.Equals("--log",StringComparison.OrdinalIgnoreCase) || arg.Equals("-l", StringComparison.OrdinalIgnoreCase))
+                {
+                    Console.WriteLine("--log/-l option detected. Will write logs into _copychanged_copy.log and _copychanged_fix.log.");
+                    doLog = true;
+                } 
+                else if (arg.Equals("--verbose",StringComparison.OrdinalIgnoreCase) || arg.Equals("-v", StringComparison.OrdinalIgnoreCase))
+                {
+                    Console.WriteLine("--verbose/-v option detected. Will write more details into _copychanged_copy.log (if --log is specified).");
+                    verboseLog = true;
                 } else if(folder1 == null)
                 {
                     folder1 = arg;
@@ -91,22 +108,100 @@ namespace copychanged
                 Console.WriteLine($"{fileToFix.to}");
             }
 
+            DateTime now = DateTime.Now;
+
+            if (doLog)
+            {
+                StringBuilder sb = new StringBuilder();
+                string header = $"[{now.ToString("yyyy-MM-dd HH:mm:ss")}]";
+
+                sb.AppendLine();
+                sb.AppendLine();
+                sb.AppendLine(header);
+                sb.AppendLine();
+                sb.AppendLine("Files to be fixed:");
+                UInt64 totalCount = (UInt64)filesToFix.Count;
+                UInt64 totalSize = 0;
+                UInt64 totalSizeDest = 0;
+                if (filesToFix.Count > 0)
+                {
+                    foreach (FileToCopy fileToFix in filesToFix)
+                    {
+                        totalSize += fileToFix.size;
+                        totalSizeDest += fileToFix.sizeTarget;
+                        string size1 = fileToFix.size.ToString("#,##0");
+                        string size2 = fileToFix.sizeTarget.ToString("#,##0");
+                        sb.AppendLine($"{fileToFix.from} ({size1} bytes) --> {fileToFix.to} ({size2} bytes)");
+                    }
+                }
+                else
+                {
+                    sb.AppendLine("None.");
+                }
+                string sizeTotal = totalSize.ToString("#,##0");
+                string sizeTotalDest = totalSizeDest.ToString("#,##0");
+                sb.AppendLine();
+                sb.AppendLine($"Total: {totalCount} files totaling {sizeTotal} bytes ({sizeTotalDest} bytes of target files at this time).");
+                sb.AppendLine();
+                File.AppendAllText("_copychanged_fix.log", sb.ToString());
+                sb.Clear();
+
+                sb.AppendLine();
+                sb.AppendLine();
+                sb.AppendLine(header);
+                sb.AppendLine();
+                sb.AppendLine("Files to be copied:");
+                totalCount = (UInt64)filesToCopy.Count;
+                totalSize = 0;
+                if (filesToCopy.Count > 0)
+                {
+                    foreach (FileToCopy fileToCopy in filesToCopy)
+                    {
+                        totalSize += fileToCopy.size;
+                        if (verboseLog)
+                        {
+                            string size1 = fileToCopy.size.ToString("#,##0");
+                            sb.AppendLine($"{fileToCopy.from} ({size1} bytes) --> {fileToCopy.to} (doesn't exist)");
+                        }
+                    }
+                }
+                else
+                {
+                    sb.AppendLine("None.");
+                }
+                sizeTotal = totalSize.ToString("#,##0");
+                sb.AppendLine();
+                sb.AppendLine($"Total: {totalCount} files totaling {sizeTotal} bytes.");
+                sb.AppendLine();
+                File.AppendAllText("_copychanged_copy.log", sb.ToString());
+                sb.Clear();
+            }
+
             CancellationTokenSource cts = new CancellationTokenSource();
 
-            if (doChanged)
+            Int64 totalSuccess = 0;
+            Int64 totalFail = 0;
+            if (doChanged && filesToFix.Count > 0)
             {
                 Console.WriteLine("Overwriting changed files now"); 
                 foreach (FileToCopy fileToFix in filesToFix)
                 {
                     Console.Write($"{fileToFix.to}...");
+                    int currentAttempt = 0;
                     try
                     {
                         bool different = true;
-                        int currentAttempt = 0;
                         while (different)
                         {
-                            File.Delete(fileToFix.to);
+                            if (File.Exists(fileToFix.to))
+                            {
+                                File.Delete(fileToFix.to);
+                            } else
+                            {
+                                Console.Write($"WEIRD, target file gone? Copying...");
+                            }
                             File.Copy(fileToFix.from, fileToFix.to);
+                            Console.Write($"Copied, verifying...");
                             different = !FilesAreSame(fileToFix.from, fileToFix.to,cts.Token);
                             if (different)
                             {
@@ -115,15 +210,85 @@ namespace copychanged
                             }
                             else
                             {
+                                totalSuccess++;
+                                if (doLog && (verboseLog || currentAttempt > 0))
+                                {
+                                    File.AppendAllText("_copychanged_fix.log", $"{fileToFix.to} successfully updated ({currentAttempt} retries)\n");
+                                }
                                 Console.Write($"File successfully updated.");
                             }
                         }
                     } catch(Exception ex)
                     {
-                        Console.Write($"Error happened when trying to overwrite.");
+                        totalFail++;
+                        Console.Write($"Error happened when trying to overwrite. {ex.ToString()}");
+                        if (doLog)
+                        {
+                            File.AppendAllText("_copychanged_fix.log", $"{fileToFix.to} failed to update ({currentAttempt} retries):\n{ex.ToString()}\n");
+                        }
                     }
                     Console.WriteLine();
                 }
+            }
+            if (doLog)
+            {
+                File.AppendAllText("_copychanged_fix.log", $"\n{totalSuccess} files successfully updated, {totalFail} failed.\n");
+            }
+
+            totalSuccess = 0;
+            totalFail = 0;
+            if (doNew && filesToCopy.Count > 0)
+            {
+                Console.WriteLine("Copying new files now"); 
+                foreach (FileToCopy fileToCopy in filesToCopy)
+                {
+                    Console.Write($"{fileToCopy.to}...");
+                    int currentAttempt = 0;
+                    try
+                    {
+                        bool different = true;
+                        while (different)
+                        {
+                            if (File.Exists(fileToCopy.to))
+                            {
+                                File.Delete(fileToCopy.to);
+                            } else if (currentAttempt != 0)
+                            {
+                                Console.Write($"WEIRD, target file gone? Copying again...");
+                            }
+                            File.Copy(fileToCopy.from, fileToCopy.to);
+                            Console.Write($"Copied, verifying...");
+                            different = !FilesAreSame(fileToCopy.from, fileToCopy.to,cts.Token);
+                            if (different)
+                            {
+                                currentAttempt++;
+                                Console.Write($"Copied file is STILL different. Retry {currentAttempt}....");
+                            }
+                            else
+                            {
+                                totalSuccess++;
+                                if (doLog && (verboseLog || currentAttempt > 0))
+                                {
+                                    File.AppendAllText("_copychanged_copy.log", $"{fileToCopy.to} successfully copied ({currentAttempt} retries)\n");
+                                }
+                                Console.Write($"File successfully copied.");
+                            }
+                        }
+                    } catch(Exception ex)
+                    {
+                        totalFail++;
+                        if (doLog)
+                        {
+                            File.AppendAllText("_copychanged_copy.log", $"{fileToCopy.to} failed to copy ({currentAttempt} retries):\n{ex.ToString()}\n");
+                        }
+                        Console.Write($"Error happened when trying to copy file. {ex.ToString()}");
+                    }
+                    Console.WriteLine();
+                }
+            }
+            if (doLog)
+            {
+                File.AppendAllText("_copychanged_copy.log", $"\n{totalSuccess} files successfully copied, {totalFail} failed.\n");
             }
 
 
@@ -168,6 +333,7 @@ namespace copychanged
             }
         }
 
+        // TODO more try here?
         static void DoFolderRecursive(string basePathReference, string basePathDestination, string reference, List<FileToCopy> filesToCopy, List<FileToCopy> filesToFix, ref UInt64 totalCompared, Stopwatch sw)
         {
             if (!Directory.Exists(reference))
@@ -184,11 +350,15 @@ namespace copychanged
                 string targetFile = Path.Combine(basePathDestination, Path.GetRelativePath(basePathReference, file));
                 if (!File.Exists(targetFile))
                 {
+                    FileInfo info = new FileInfo(file);
+
                     //Console.WriteLine($"target file {targetFile} doesn't exist. must copy.");
                     filesToCopy.Add(new FileToCopy() { 
                         from=file,
                         to=targetFile,
-                        existsAndDifferent = false
+                        existsAndDifferent = false,
+                        size = (UInt64)info.Length,
+                        sizeTarget = 0
                     });
                     continue;
                 }
@@ -196,11 +366,13 @@ namespace copychanged
                 {
                     bool same;
                     UInt64 length1 = 0;
+                    UInt64 length2 = 0;
                     using (FileStream fs1 = File.OpenRead(file))
                     {
                         length1 = (UInt64)fs1.Length;
                         using (FileStream fs2 = File.OpenRead(targetFile))
                         {
+                            length2 = (UInt64)fs2.Length;
                             same = VectorizedComparer.Same(fs1, fs2, cts.Token, default, true);
                         }
                     }
@@ -217,7 +389,9 @@ namespace copychanged
                         {
                             from = file,
                             to = targetFile,
-                            existsAndDifferent = true
+                            existsAndDifferent = true,
+                            size = length1,
+                            sizeTarget = length2
                         });
                     }
